pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "resource-librarian"
version = "0.1.0"
description = "A library to help create, maintain and index a digital library of electronic book, youtube transcripts and summaries of them"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Kenny Wilson"}
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
    "pymupdf>=1.23.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "google-api-python-client>=2.0.0",
    "youtube-transcript-api>=0.6.0",
]

[project.scripts]
rl = "resourcelibrarian.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.hatch.build.targets.wheel]
packages = ["src/resourcelibrarian"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]
